Move functions CreateConceptMap and SaveConceptMap from service to converter
